+++ src/RenderEngine/scene/model/UV.cpp
@@ -5,7 +5,7 @@ UV::UV()
 {
 }
 
-UV::UV(double _u, double _v)
+UV::UV(float _u, float _v)
 {
     u = _u;
     v = _v;
@@ -17,5 +17,5 @@ UV::~UV()
 
 vec2 UV::to_vec() const
 {
-    return vec2({static_cast<float>(u), static_cast<float>(v)});
+    return vec2({u, v});
 }
